pyproject.toml
[project]
name = "livekit-oubound-call-dispatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "livekit-api>=1.1.0",
    "pydantic-settings>=2.13.1",
]
